+++ packages/datadog-core/test/utils/src/parse-tags.spec.js
@@ -1,0 +1,23 @@
+'use strict'
+
+require('../../../../dd-trace/test/setup/tap')
+
+const parseTags = require('../../../src/utils/src/parse-tags')
+
+describe('parseTags', () => {
+  it('should parse tags to object', () => {
+    const obj = {
+      'a.0.a': 'foo',
+      'a.0.b': 'bar',
+      'a.1.a': 'baz'
+    }
+
+    expect(parseTags(obj)).to.deep.equal({
+      a: [{ a: 'foo', b: 'bar' }, { a: 'baz' }]
+    })
+  })
+
+  it('should work with empty object', () => {
+    expect(parseTags({})).to.deep.equal({})
+  })
+})
