+++ packages/dd-trace/src/plugins/outbound.js
@@ -7,7 +7,7 @@ const {
   PEER_SERVICE_REMAP_KEY
 } = require('../constants')
 const TracingPlugin = require('./tracing')
-const { exitTag } = require('../../../datadog-code-origin')
+const { exitTags } = require('../../../datadog-code-origin')
 
 const COMMON_PEER_SVC_SOURCE_TAGS = [
   'net.peer.name',
@@ -29,7 +29,7 @@ class OutboundPlugin extends TracingPlugin {
   startSpan (...args) {
     const span = super.startSpan(...args)
     if (this._tracerConfig.codeOriginForSpans.enabled) {
-      span.addTags(exitTag(this.startSpan))
+      span.addTags(exitTags(this.startSpan))
     }
     return span
   }
